refactored components and static assets
