+++ portal/src/assets/images/closed.js
