+++ portal/src/assets/images/favicon.ico
