+++ portal/src/assets/images/logo@2x.png
