+++ portal/src/assets/images/night.png
