+++ portal/src/assets/images/opened.js
