+++ portal/src/assets/images/pwa-512.png
