+++ portal/src/components/docsComponent/draft/Draft.js
