+++ portal/src/components/index.js
@@ -2,4 +2,4 @@ export * from './theme';
 export mdxComponents from './mdxComponents';
 export ThemeProvider from './theme/themeProvider';
 export Layout from './layout';
-export Link from './link';
+export Link from './staticComponents/link';
