+++ portal/src/components/rightSidebar/index.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { StaticQuery, graphql } from 'gatsby';
 
 // import Link from './link';
-import config from '../../config';
-import { Sidebar, ListItem } from './styles/Sidebar';
+import config from '../../../config';
+import { Sidebar, ListItem } from './Sidebar';
 
 const SidebarLayout = ({ location }) => (
   <StaticQuery
