+++ portal/src/components/staticComponents/link.js
